+++ RELEASENOTES.md
@@ -71,6 +71,12 @@
   show how to render video to a `GLSurfaceView` while applying a GL shader.
   ([#6920](https://github.com/google/ExoPlayer/issues/6920)).
 
+### 2.11.3 (2020-02-19) ###
+
+* DRM: Fix issue switching from protected content that uses a 16-byte
+  initialization vector to one that uses an 8-byte initialization vector
+  ([#6982](https://github.com/google/ExoPlayer/issues/6982)).
+
 ### 2.11.2 (2020-02-13) ###
 
 * Add Java FLAC extractor